renamed blotter files, fixed displayed rows status
