+++ src/client/src/ui/blotter/blotter.scss
