+++ src/client/src/ui/blotter/index.ts
@@ -1,4 +1,4 @@
-export { default as AgGridBlotter } from './AgGridBlotter'
+export { default as Blotter } from './Blotter'
 export { default as BlotterContainer } from './BlotterContainer'
 export { default as blotterReducer }  from './reducer'
 export { default as blotterEpic } from './epics'
